fix: Domain list types
