+++ src/domains/interfaces/create-domain-options.interface.ts
@@ -1,6 +1,5 @@
 import { PostOptions } from '../../common/interfaces';
-
-export type DomainRegion = 'us-east-1' | 'eu-west-1' | 'sa-east-1';
+import { DomainRegion, DomainStatus } from './domain';
 
 export interface CreateDomainOptions {
   name: string;
@@ -8,23 +7,6 @@ export interface CreateDomainOptions {
 }
 
 export interface CreateDomainRequestOptions extends PostOptions {}
-
-export type DomainNameservers =
-  | 'Amazon Route 53'
-  | 'Cloudflare'
-  | 'Digital Ocean'
-  | 'GoDaddy'
-  | 'Google Domains'
-  | 'Namecheap'
-  | 'Unidentified'
-  | 'Vercel';
-
-export type DomainStatus =
-  | 'pending'
-  | 'verified'
-  | 'failed'
-  | 'temporary_failure'
-  | 'not_started';
 
 export interface DomainSpfRecord {
   record: 'SPF';
