+++ src/domains/interfaces/index.ts
@@ -1,1 +1,2 @@
 export * from './create-domain-options.interface';
+export * from './list-domains.interface';
