+++ src/domains/interfaces/list-domains.interface.ts
@@ -1,0 +1,11 @@
+import { DomainRegion, DomainStatus } from './domain';
+
+export type ListDomainsResponse = {
+  data: {
+    name: string;
+    id: string;
+    created_at: string;
+    region: DomainRegion;
+    status: DomainStatus;
+  }[];
+};
